moved Create Item dialog box to be a separate component
